+++ packages/common/src/ee/AiAgent/schemas/tools/index.ts
@@ -3,6 +3,7 @@ export * from './toolFindChartsArgs';
 export * from './toolFindDashboardsArgs';
 export * from './toolFindExploresArgs';
 export * from './toolFindFieldsArgs';
+export * from './toolImproveContext';
 export * from './toolRunMetricQueryArgs';
 export * from './toolSearchFieldValuesArgs';
 export * from './toolTableVizArgs';